allow prewrite hook to add more files to the tree that will also be passed through the prewrite phase
this feels really weird to me, no matter how I look at it. I think the best way to handle this is to
create a brand new hook, but I can't think of any other use-cases or even what to name it. I'll revisit
this after I can get the sourcemaps plugin working. (that's the main reason I'm adding this)